Update 1.0.6
Code wise - totally minuscule update. However technically it was unexpected battle.

I believe that this is a story every programmer faced in his time at least once. U think "I have spare 15 minutes, let's add something easy! Just for sake of it being there waiting for me to take it serious.". All of a sudden it's 11 PM and you started at 7 PM. You are facing a cascade of something not supporting something else... And you wanted to add a simple fricking icon to ease your mind from dictionaries, loops and recurrence but all of a sudden you had to perform a mental backflip to add a stupid icon... Over 4 hours of your life are long time gone... but there it is. An icon. Curtains!